backend/requirements.txt
Django==3.1.1
django-cors-headers==3.2.1
djangorestframework==3.11.1
